# Optimize add_interconnects() for faster switch sync