If we have about 2000 interconnects that we want to put into a switch config adding them takes a long time. To avoid asking the DB each time, we now build up a list of (physnet, network_id) tuples and query all the segments out in one biiig query.

A short non-representative benchmark has shown that this code now takes 9.6s instead of 42s. The remaining time is spent in building up the tuples (~6s), getting the segments (~600ms) and adding the data to the config (~3s).